+++ config.js
@@ -1,0 +1,15 @@
+const config = {
+  /*
+ * - Insert each Twitter credential inside it's equivalent quote.
+ */
+  consumer_key: '',
+  consumer_secret: '',
+  access_token: '',
+  access_token_secret: '',
+  /*
+ * - Change 'mybot' with your bot account '@'.
+ * - Example: if your account '@' is '@twitter_bot', you should write only 'twitter_bot'
+ */
+  screenName: 'mybot'
+};
+exports.config = config;
